requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
Pillow>=8.3.0
scikit-learn>=0.24.0
tqdm>=4.62.0
requests>=2.26.0
psutil>=5.8.0